changing fips codes to char data type

+++ src/analysis/import/import_osm.sh
@@ -152,7 +152,7 @@ echo 'START: Importing State Default Speed Table'
 psql -h $NB_POSTGRESQL_HOST -U $NB_POSTGRESQL_USER -d $NB_POSTGRESQL_DB \
      -c "CREATE TABLE IF NOT EXISTS \"state_speed\" (
             state char(2),
-            fips_code_state smallint,
+            fips_code_state char(2),
             speed smallint
         );"
 
@@ -161,11 +161,11 @@ STATE_SPEED_FILENAME="state_fips_speed"
 STATE_SPEED_DOWNLOAD="/data/${STATE_SPEED_FILENAME}.csv"
 wget -nv -O "${STATE_SPEED_DOWNLOAD}" "https://s3.amazonaws.com/pfb-public-documents/${STATE_SPEED_FILENAME}.csv"
 psql -h $NB_POSTGRESQL_HOST -U $NB_POSTGRESQL_USER -d $NB_POSTGRESQL_DB \
-   -c "\copy state_speed FROM ${STATE_SPEED_DOWNLOAD} delimiter ',' csv header"
+     -c "\copy state_speed FROM ${STATE_SPEED_DOWNLOAD} delimiter ',' csv header"
 
 # Set default residential speed for state
 STATE_DEFAULT=$( psql -h $NB_POSTGRESQL_HOST -U $NB_POSTGRESQL_USER -d $NB_POSTGRESQL_DB \
-      -t -c "SELECT state_speed.speed FROM state_speed WHERE state_speed.fips_code_state = ${PFB_STATE_FIPS}" )
+      -t -c "SELECT state_speed.speed FROM state_speed WHERE state_speed.fips_code_state = '${PFB_STATE_FIPS}'" )
 echo "DONE: Importing state default residential speed"
 
 # Create table for city residential speeds
@@ -174,26 +174,33 @@ psql -h $NB_POSTGRESQL_HOST -U $NB_POSTGRESQL_USER -d $NB_POSTGRESQL_DB \
      -c "CREATE TABLE IF NOT EXISTS \"city_speed\" (
             city varchar,
             state char(2),
-            fips_code_city integer,
+            fips_code_city char(7),
             speed smallint
         );"
 
 # Import city residential speeds file
 CITY_SPEED_FILENAME="city_fips_speed"
 CITY_SPEED_DOWNLOAD="/data/${CITY_SPEED_FILENAME}.csv"
-wget -nv -O "${CITY_SPEED_DOWNLOAD}" "https://s3.amazonaws.com/pfb-public-documents/${CITY_SPEED_FILENAME}.csv"
+if [ -f "/data/${CITY_SPEED_FILENAME}.csv" ]; then
+	echo "Using local city speed file"
+else
+	wget -nv -O "${CITY_SPEED_DOWNLOAD}" "https://s3.amazonaws.com/pfb-public-documents/${CITY_SPEED_FILENAME}.csv"
+fi
+
 psql -h $NB_POSTGRESQL_HOST -U $NB_POSTGRESQL_USER -d $NB_POSTGRESQL_DB \
-   -c "\copy city_speed FROM ${CITY_SPEED_DOWNLOAD} delimiter ',' csv header"
+     -c "\copy city_speed FROM ${CITY_SPEED_DOWNLOAD} delimiter ',' csv header"
 
 # Set default residential speed for city
 CITY_DEFAULT=$( psql -h $NB_POSTGRESQL_HOST -U $NB_POSTGRESQL_USER -d $NB_POSTGRESQL_DB \
-      -t -c "SELECT city_speed.speed FROM city_speed WHERE city_speed.fips_code_city = ${PFB_CITY_FIPS}" )
+      -t -c "SELECT city_speed.speed FROM city_speed WHERE city_speed.fips_code_city = '${PFB_CITY_FIPS}'" )
 # Check if no value for city default, if so set to NULL
-echo "The city residential default speed is ${CITY_DEFAULT}."
+
 if [[ -z "$CITY_DEFAULT" ]];
 then
-  echo "No default residential speed in city"
-  CITY_DEFAULT=NULL
+    echo "No default residential speed in city"
+    CITY_DEFAULT=NULL
+else
+    echo "The city residential default speed is ${CITY_DEFAULT}."
 fi
 echo "DONE: Importing city default residential speed"
